Resolvendo testes redundantes com nomes que representam sua classe de equivalencia

+++ src/capitulo6/Calculadora.java
@@ -2,7 +2,7 @@ package capitulo6;
 
 public class Calculadora {
 
-	public double soma(int a, int b) {
+	public double soma(double a, double b) {
 		return a + b;
 	}
